feat(pci.ai-dashboard): review application architecture
