+++ packages/manager/apps/pci-ai-dashboard/public/translations/pci-ai-dashboard/Messages_fr_FR.json
