+++ packages/manager/apps/pci-ai-dashboard/public/translations/regions/Messages_fr_FR.json
