+++ packages/manager/apps/pci-ai-dashboard/src/data/api/projects.ts
@@ -1,5 +1,5 @@
 import { apiClient } from '@ovh-ux/manager-core-api';
-import { PciProject } from '@/interfaces/project';
+import { PciProject } from '@/types/project';
 
 export const getProject = async (projectId: string): Promise<PciProject> => {
   const response = await apiClient.v6.get(`/cloud/project/${projectId}`);
